Issue #414 preflight: Move oc caching to common code
Now that 69f693fe introduced a file for preflight checks common to all
platforms, we can move the preflight check that the oc binary is present
to that file rather than duplicating it between all the platforms.

This fixes https://github.com/code-ready/crc/issues/380 and
https://github.com/code-ready/crc/issues/414
