+++ pkg/crc/preflight/preflight_checks_linux.go
@@ -18,7 +18,6 @@ import (
 	"github.com/code-ready/crc/pkg/crc/constants"
 	"github.com/code-ready/crc/pkg/crc/logging"
 	"github.com/code-ready/crc/pkg/crc/machine/libvirt"
-	"github.com/code-ready/crc/pkg/crc/oc"
 	"github.com/code-ready/crc/pkg/crc/systemd"
 
 	crcos "github.com/code-ready/crc/pkg/os"
@@ -469,24 +468,5 @@ func fixCrcNetworkManagerConfig() (bool, error) {
 	}
 
 	logging.Debug("NetworkManager configuration fixed")
-	return true, nil
-}
-
-// Check if oc binary is cached or not
-func checkOcBinaryCached() (bool, error) {
-	oc := oc.OcCached{}
-	if !oc.IsCached() {
-		return false, errors.New("oc binary is not cached.")
-	}
-	logging.Debug("oc binary already cached")
-	return true, nil
-}
-
-func fixOcBinaryCached() (bool, error) {
-	oc := oc.OcCached{}
-	if err := oc.EnsureIsCached(); err != nil {
-		return false, fmt.Errorf("Not able to download oc %v", err)
-	}
-	logging.Debug("oc binary cached")
 	return true, nil
 }
